Update messages about downloading missing wordlist
Instead of just asking `download default wordlist? [y/n]', the URL which
will be downloaded is shown, to enable users to easily check out the
file that will be downloaded and dissuade any sort of sneaky
URL-changing shenanigans.

The instructions on how to download the default wordlist manually
provided in the event curl is not installed have also been updated to be
more explicit and helpful.